remove contingencies from security analysis
