+++ src/main/java/com/powsybl/openloadflow/sa/OpenSecurityAnalysis.java
@@ -26,6 +26,7 @@ import com.powsybl.openloadflow.ac.outerloop.AcloadFlowEngine;
 import com.powsybl.openloadflow.equations.*;
 import com.powsybl.openloadflow.graph.GraphDecrementalConnectivity;
 import com.powsybl.openloadflow.network.*;
+import com.powsybl.openloadflow.network.impl.LfContingency;
 import com.powsybl.openloadflow.network.util.ActivePowerDistribution;
 import com.powsybl.security.*;
 import com.powsybl.security.interceptors.SecurityAnalysisInterceptor;
@@ -305,7 +306,7 @@ public class OpenSecurityAnalysis implements SecurityAnalysis {
         List<Equation> deactivatedEquations = new ArrayList<>();
         List<EquationTerm> deactivatedEquationTerms = new ArrayList<>();
 
-        deactivateEquations(lfContingency, engine.getEquationSystem(), deactivatedEquations, deactivatedEquationTerms);
+        LfContingency.deactivateEquations(lfContingency, engine.getEquationSystem(), deactivatedEquations, deactivatedEquationTerms);
 
         // restart LF on post contingency equation system
         engine.getParameters().setVoltageInitializer(new PreviousValueVoltageInitializer());
@@ -319,73 +320,13 @@ public class OpenSecurityAnalysis implements SecurityAnalysis {
                 postContingencyLimitViolations);
         }
 
-        reactivateEquations(deactivatedEquations, deactivatedEquationTerms);
+        LfContingency.reactivateEquations(deactivatedEquations, deactivatedEquationTerms);
 
         stopwatch.stop();
         LOGGER.info("Post contingency '{}' simulation done in {} ms", lfContingency.getContingency().getId(),
                 stopwatch.elapsed(TimeUnit.MILLISECONDS));
 
         return new PostContingencyResult(lfContingency.getContingency(), postContingencyComputationOk, postContingencyLimitViolations);
-    }
-
-    // TODO: extract it properly
-    public static void deactivateEquations(LfContingency lfContingency, EquationSystem equationSystem, List<Equation> deactivatedEquations, List<EquationTerm> deactivatedEquationTerms) {
-        for (LfBranch branch : lfContingency.getBranches()) {
-            LOGGER.trace("Remove equations and equations terms related to branch '{}'", branch.getId());
-
-            // deactivate all equations related to a branch
-            for (Equation equation : equationSystem.getEquations(SubjectType.BRANCH, branch.getNum())) {
-                if (equation.isActive()) {
-                    equation.setActive(false);
-                    deactivatedEquations.add(equation);
-                }
-            }
-
-            // deactivate all equation terms related to a branch
-            for (EquationTerm equationTerm : equationSystem.getEquationTerms(SubjectType.BRANCH, branch.getNum())) {
-                if (equationTerm.isActive()) {
-                    equationTerm.setActive(false);
-                    deactivatedEquationTerms.add(equationTerm);
-                }
-            }
-        }
-
-        for (LfBus bus : lfContingency.getBuses()) {
-            LOGGER.trace("Remove equations and equation terms related to bus '{}'", bus.getId());
-
-            // deactivate all equations related to a bus
-            for (Equation equation : equationSystem.getEquations(SubjectType.BUS, bus.getNum())) {
-                if (equation.isActive()) {
-                    equation.setActive(false);
-                    deactivatedEquations.add(equation);
-                }
-            }
-
-            // deactivate all equation terms related to a bus
-            for (EquationTerm equationTerm : equationSystem.getEquationTerms(SubjectType.BUS, bus.getNum())) {
-                if (equationTerm.isActive()) {
-                    equationTerm.setActive(false);
-                    deactivatedEquationTerms.add(equationTerm);
-                }
-            }
-        }
-    }
-
-    // TODO: extract it properly
-    public static void reactivateEquations(List<Equation> deactivatedEquations, List<EquationTerm> deactivatedEquationTerms) {
-        // restore deactivated equations and equations terms from previous contingency
-        if (!deactivatedEquations.isEmpty()) {
-            for (Equation equation : deactivatedEquations) {
-                equation.setActive(true);
-            }
-            deactivatedEquations.clear();
-        }
-        if (!deactivatedEquationTerms.isEmpty()) {
-            for (EquationTerm equationTerm : deactivatedEquationTerms) {
-                equationTerm.setActive(true);
-            }
-            deactivatedEquationTerms.clear();
-        }
     }
 
     List<LfContingency> createContingencies(List<ContingencyContext> contingencyContexts, LfNetwork network) {
